+++ src/data/datacube/aggregator.h
@@ -24,8 +24,7 @@ public:
 	explicit Aggregator(Type type);
 	Aggregator &add(double);
 	Aggregator &add(const Aggregator &);
-	[[nodiscard]] explicit operator bool() const;
-	[[nodiscard]] double operator*() const;
+	[[nodiscard]] explicit operator double() const;
 	[[nodiscard]] bool isEmpty() const;
 
 private: