+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 - Line chart connector circles color was not contained the alpha channel.
 - Line chart draws was overwrite the event's settings.
 - Legend title outerRect was not properly calculated.
-- Fixed stacked empty min/max/mean aggregated values.
+- Fixed stacked empty min/max aggregated values.
 
 ### Added